+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.3
+
+* Added ``blockUntil`` function to the ``WaitChainBuilder``.
+* ``SpriteBatch`` render performance increased.
+
 ## 1.0.2
 
 * Added the ``WaitChainBuilder`` class to make ``WaitEvents`` easier to use. It allows you to create game cinematics more easily.